requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Medical Imaging
monai>=1.3.0
nibabel>=5.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.2.0

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.65.0

# Logging
tensorboard>=2.13.0

# Image Processing
SimpleITK>=2.2.0

# Utilities
python-dotenv>=1.0.0
